+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/handler/repository/RoutedRepository.java
@@ -19,9 +19,9 @@ import java.util.*;
  * @since 1.0.0
  */
 @Component
-public class RoutedRepository implements Repository {
+public class RoutedRepository extends AbstractRepository {
 
-    public static final String SUPPORT_TYPE = "routed";
+    public static final String REPOSITORY_TYPE = "routed";
 
     @SuppressWarnings("FieldMayBeFinal")
     @Autowired(required = false)
@@ -34,12 +34,16 @@ public class RoutedRepository implements Repository {
 
     private final Map<String, Repository> repositoryMap = new HashMap<>();
 
+    public RoutedRepository() {
+        super(REPOSITORY_TYPE);
+    }
+
     @PostConstruct
     public void init() {
         StringTokenizer stringTokenizer = new StringTokenizer(routedTypes, ",");
         while (stringTokenizer.hasMoreTokens()) {
             String repositoryType = stringTokenizer.nextToken();
-            if (Objects.equals(SUPPORT_TYPE, repositoryType)) {
+            if (Objects.equals(REPOSITORY_TYPE, repositoryType)) {
                 throw new IllegalArgumentException(
                         "RoutedRepository 不能引用自身，请修改 repository.routed.supported_types 属性");
             }
@@ -48,11 +52,6 @@ public class RoutedRepository implements Repository {
                             "，请修改 repository.routed.supported_types 属性"));
             repositoryMap.put(repositoryType, repository);
         }
-    }
-
-    @Override
-    public boolean supportType(String type) {
-        return Objects.equals(SUPPORT_TYPE, type);
     }
 
     @Override
@@ -68,5 +67,31 @@ public class RoutedRepository implements Repository {
         } catch (Exception e) {
             throw new RepositoryException(e);
         }
+    }
+
+    @Override
+    public void putData(String category, Object[] args) throws RepositoryException {
+        try {
+            String[] split = category.split(delimiter, 2);
+            if (!repositoryMap.containsKey(split[0])) {
+                throw new RepositoryException("未知的 repository 类型: " + split[0]);
+            }
+            repositoryMap.get(split[0]).putData(split[1], args);
+        } catch (RepositoryException e) {
+            throw e;
+        } catch (Exception e) {
+            throw new RepositoryException(e);
+        }
+    }
+
+    @Override
+    public String toString() {
+        return "RoutedRepository{" +
+                "repositories=" + repositories +
+                ", routedTypes='" + routedTypes + '\'' +
+                ", delimiter='" + delimiter + '\'' +
+                ", repositoryMap=" + repositoryMap +
+                ", repositoryType='" + repositoryType + '\'' +
+                '}';
     }
 }