+++ judge-stack/src/main/java/com/dwarfeng/judge/stack/handler/RepositoryHandler.java
@@ -20,4 +20,13 @@ public interface RepositoryHandler extends Handler {
      * @throws HandlerException 处理器异常。
      */
     Object getData(String category, Object... args) throws HandlerException;
+
+    /**
+     * 向仓库中推送数据。
+     *
+     * @param category 数据的分类。
+     * @param args     数据的详细参数。
+     * @throws HandlerException 处理器异常。
+     */
+    void putData(String category, Object... args) throws HandlerException;
 }